% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/log_lik.R
\name{log_lik}
\alias{log_lik}
\title{Extract pointwise log-likelihood from bipartite model}
\usage{
log_lik(fit, ...)
}
\arguments{
\item{fit}{Fitted model (from \code{\link[=fit_model]{fit_model()}})}

\item{...}{Additional arguments passed to \code{\link[cmdstanr:as_draws.CmdStanMCMC]{cmdstanr::as_draws()}}.}
}
\value{
An object of class \code{draws_matrix}. See \code{\link[posterior:draws_matrix]{posterior::draws_matrix()}} for details.
}
\description{
Extract pointwise log-likelihood from bipartite model
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
data(web)
dt <- prepare_data(mat = web, plant_effort = rep(20, nrow(web)))
fit <- fit_model(dt)
ll <- log_lik(fit)

\dontshow{\}) # examplesIf}
}
